mkochman-szpila-865f30290cad/requirements.txt
Mezzanine==3.1.10
django-rosetta
pep8
django-modeltranslation
pyflakes
django-compressor